+++ src/components/profile.js
@@ -1,7 +1,10 @@
 export const createHeaderProfileComponent = (rank) => {
-  return `
-<section class="header__profile profile">
-  <p class="profile__rating">${rank}</p>
-  <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-</section>`;
+  const isHasRank = rank !== `neither`;
+  
+  return `${isHasRank ? `
+    <section class="header__profile profile">
+      <p class="profile__rating">${rank}</p>
+      <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
+    </section>`
+  : ``}`
 };
