+++ src/mock/most-commented.js
@@ -3,7 +3,7 @@ const generateMostCommented = (films) => {
     arr.sort((a, b) => (a.numberOfComments > b.numberOfComments ? 1 : -1)).reverse();
   };
   sortByСomment(films);
-  return films.slice(0, 2);
+  return films[0].numberOfComments === 0 ? false :  films.slice(0, 2);
 };
 
 export {generateMostCommented};
