+++ src/mock/top-rated.js
@@ -3,7 +3,7 @@ const generateTopRated = (films) => {
     arr.sort((a, b) => (a.rating > b.rating ? 1 : -1)).reverse();
   };
   sortByRating(films);
-  return films.slice(0, 2);
+  return films[0].rating === 0 ? false :  films.slice(0, 2);
 };
 
 export {generateTopRated};
